refactor(hnsw): split hnsw.rs into hnsw/ directory + fix flaky test
Split 1150-line monolith into focused modules:
- hnsw/mod.rs: types, constants, VectorIndex impl, Send+Sync tests
- hnsw/build.rs: build(), build_batched() + build tests
- hnsw/search.rs: search() implementation
- hnsw/persist.rs: save(), load(), checksums, try_load() + persist tests
- hnsw/safety.rs: LoadedHnsw safety tests (lifecycle, threading, layout)

Fix flaky test_loaded_index_multiple_searches: replaced sin-based
embeddings (too similar in cosine space) with one-hot-like vectors
that have clear separation (~0.01 cosine similarity between seeds).

No behavior changes. All 258 tests pass, flaky test verified 5x stable.
